Dev experience: missing client registration debug
When circumventing the FromRobot(r) helpers, like one may do if they
are new to the RDK, the likely result of trying to address any one of
r.ResourceNames is going to lead to the ErrMissingClientRegistration.
If said soul is familiar with Go own flavor of DI package registration
it's easy to find the x/register packages (and probably the helpers in
the process). However a slight nudge never hurt anyone.